phase-B-R5b: thermal spinor result — holonomy frozen, spinor invisible
R5b (спинор κ=1, отжиг T0=1.0, decay∈{0.999,0.9995,0.9999}, блочная сходимость
после T<1e-4, 817с):
- Плотность кинков = P(h=−1) = 0.0000 при ВСЕХ decay (даже 0.9999, T_fin=3.7e-6).
  Пре-рег (a) «отжиг населяет кинки» ОПРОВЕРГНУТ: энергия кинка k_e·arccos²≈155…620
  (k_e=63) ≫ T0=1.0 ⇒ ℤ₂-сектор не населяется термически.
- Голономия h=+1 у всех лент ⇒ E(θ|h−) не определена, M2≡M1 (нет антиферро).
- (b) E ферро ✅; амплитуда РАСТЁТ с замедлением отжига (0.49→0.66→0.99) — отжиг
  ведёт κ=1 к аглайн-глобальному минимуму (ступень), превью R2 §4.
- decay=0.9999 (амп 0.99) — контроли маргиналов/симметрии падают (насыщение).
Итог спинорной линии: спинор динамически невидим при T0≤1.0 (no-go R5 + заморозка h).

Также: dynamics noise.astype(q.dtype) — корректный float64-прогон (scan-carry).
margin_rot.py: тест поворота κ=0.1 (инкрементальная запись, float64 на CPU).

+++ sim/src/ribbon_sim/dynamics.py
@@ -49,7 +49,8 @@ def build_relaxer(cfg):
             T = T0 * decay ** (step0 + i).astype(jnp.float32)
             g = grad_batch(q, a, b)
             # q += −lr*grad + sqrt(2*lr*T)*noise, затем проекция на S³.
-            noise = jax.random.normal(sub, q.shape) * jnp.sqrt(2.0 * lr * T)
+            # noise приводим к dtype состояния (иначе float64-прогон ломает scan-carry).
+            noise = (jax.random.normal(sub, q.shape) * jnp.sqrt(2.0 * lr * T)).astype(q.dtype)
             q = normalize(q - lr * g + noise)
             e = jnp.mean(energy_batch(q, a, b))
             return (q, k), e
